Move Bugsnag middleware classes into separate file
