+++ bin/maze-runner
@@ -18,6 +18,7 @@ require_relative '../lib/maze/api/appium/file_manager'
 require_relative '../lib/maze/api/appium/ui_manager'
 require_relative '../lib/maze/api/cucumber/scenario'
 require_relative '../lib/maze/api/exit_code'
+require_relative '../lib/maze/bugsnag/middleware'
 require_relative '../lib/maze/bugsnag/config'
 require_relative '../lib/maze/client/bb_api_client'
 require_relative '../lib/maze/client/bb_client_utils'
